fixtures/security_lab_v2/requirements.txt
Flask>=3.0
Jinja2>=3.1
PyJWT>=2.8
dill>=0.3
jsonpickle>=3.2
lxml>=5.0
pycryptodome>=3.20
requests>=2.31
